feat: Add MCQ.ts & change main MCQ model
- interface MCQAttribute
- interface MCQReaderViewProps
- mcq_types.ts is Temporarily commented out unused code
