+++ src/model/interface/MCQ.ts
@@ -1,0 +1,14 @@
+import { Painting } from './painting';
+
+export interface MCQAttribute {
+    distractorPaintings: Painting[];
+    answerPaintings: Painting[];
+    category: string;
+    similarity: string;
+}
+
+export interface MCQReaderViewProps {
+    attribute: MCQAttribute;
+    currentAttributeIndex: number;
+    handelNextMCQ: () => void;
+}
