+++ src/types/mcq_types.ts
@@ -1,37 +1,33 @@
-// extracted types from the MCQ conglomerate
-// the first two are for the API, the last 3 are from the frontend code
+// TODO 참고로 사용후 제거 예정
 
-import { Painting } from '@/mock/data/entity/mock_painting';
+// // extracted types from the MCQ conglomerate
+// // the first two are for the API, the last 3 are from the frontend code
 
-export interface MCQSubmission {
-    mcq_id: string;
-    selected_answer: string;
-    is_correct: boolean;
-}
+// import { MockPainting } from '@/mock/data/entity/mock_painting';
 
-export interface MCQHintRequest {
-    question: string;
-    attemptedAnswers: string[];
-    remainingAnswers: string[];
-}
+// export interface MCQSubmission {
+//     mcq_id: string;
+//     selected_answer: string;
+//     is_correct: boolean;
+// }
 
-export interface MCQAttribute {
-    question: string;
-    displayPaintings: Painting[];
-    answers: Painting[];
-    selectedAnswer: number | null;
-    isFinalized: boolean;
-    id: string;
-    showHintButton?: boolean;
-}
+// export interface MCQHintRequest {
+//     question: string;
+//     attemptedAnswers: string[];
+//     remainingAnswers: string[];
+// }
 
-export interface MCQInstructorViewProps {
-    attribute: MCQAttribute;
-    updateAttributes: (attribute: Partial<MCQAttribute>) => void;
-}
+// export interface MCQAttribute {
+//     question: string;
+//     displayPaintings: MockPainting[];
+//     answers: MockPainting[];
+//     selectedAnswer: number | null;
+//     isFinalized: boolean;
+//     id: string;
+//     showHintButton?: boolean;
+// }
 
-export interface MCQReaderViewProps {
-    attribute: MCQAttribute;
-    currentAttributeIndex: number;
-    handelNextMCQ: () => void;
-}
+// export interface MCQInstructorViewProps {
+//     attribute: MCQAttribute;
+//     updateAttributes: (attribute: Partial<MCQAttribute>) => void;
+// }
